# src/swagger_documentation.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "DigitalStore API",
    "description": "Documentação da API da digital store"
  },
  "host": "localhost:8000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "JWT": {
      "description": "JWT token",
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/marcas/": {
      "get": {
        "tags": [
          "Marcas"
        ],
        "description": "Retorna lista de marcas",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna lista de marcas",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "marca_id": {
                    "type": "number",
                    "example": 1
                  },
                  "marca_nome": {
                    "type": "string",
                    "example": "Nike"
                  },
                  "criado_em": {
                    "type": "string",
                    "example": "2024-08-26 23:31:00"
                  },
                  "atualizado_em": {
                    "type": "string",
                    "example": "2024-08-26 23:31:00"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Marcas"
        ],
        "description": "Cria uma marca",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "marca_nome": {
                  "type": "string",
                  "example": "Puma"
                }
              },
              "required": [
                "marca_nome"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marca criada",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "detail": {
                  "type": "string",
                  "example": "Marca criada"
                },
                "severity": {
                  "type": "string",
                  "example": "success"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Marcas"
        ],
        "description": "Edita uma marca",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "marca_id": {
                  "type": "number",
                  "example": 1
                },
                "marca_nome": {
                  "type": "string",
                  "example": "Nike"
                }
              },
              "required": [
                "marca_id",
                "marca_nome"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marca atualizada",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "detail": {
                  "type": "string",
                  "example": "Marca atualizada"
                },
                "severity": {
                  "type": "string",
                  "example": "success"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/marcas/{id}": {
      "get": {
        "tags": [
          "Marcas"
        ],
        "description": "Retorna uma marca",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna uma marca",
            "schema": {
              "type": "object",
              "properties": {
                "marca_id": {
                  "type": "number",
                  "example": 1
                },
                "marca_nome": {
                  "type": "string",
                  "example": "Nike"
                },
                "criado_em": {
                  "type": "string",
                  "example": "2024-08-26 23:31:00"
                },
                "atualizado_em": {
                  "type": "string",
                  "example": "2024-08-26 23:31:00"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Marcas"
        ],
        "description": "Deleta uma marca",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Marca deletada",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "detail": {
                  "type": "string",
                  "example": "Marca deletada"
                },
                "severity": {
                  "type": "string",
                  "example": "success"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    }
  },
...
}